Bajaj Hackathon/requirements.txt
langchain
chromadb
streamlit
pypdf
pandas
groq
python-dotenv 